Re-order OS detection if..else statement.
This patch will re-order the if statements when checking what operating
system the user is running.

This is useful for people that already have a `notify-send` binary in
their $PATH but are running OSX. An example of such requirements is
anyone running `vagrant-notify` is required to have a `notify-send`
binary on their $PATH.

See [link](https://github.com/fgrehm/vagrant-notify#os-x).